Day 7: Error in reading from the api, spring security prevents postman to view the data after signup, but successfully updates in the database. updated AuthController, AuthenticationResponse, AuthService, AuthServiceImpl, pom, User, UserDto, And WebSecurityConfiguration for this issue, still facing the error.

+++ src/main/java/com/example/springBoot_taskManagement/dto/AuthenticationResponse.java
@@ -8,7 +8,7 @@ public class AuthenticationResponse {
 
     private String jwt;
 
-    private int userId;
+    private Long userId;
 
     private UserRole userRole;
 
@@ -17,7 +17,7 @@ public class AuthenticationResponse {
         this.jwt = jwt;
     }
 
-    public void setUserId(int userId) {
+    public void setUserId(Long userId) {
         this.userId = userId;
     }
