+++ src/main/java/com/example/springBoot_taskManagement/dto/UserDto.java
@@ -6,7 +6,7 @@ import lombok.Data;
 @Data
 public class UserDto {
 
-    private int id;
+    private Long id;
 
     private String name;
 
@@ -32,7 +32,7 @@ public class UserDto {
         this.name = name;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
 }
